[WEBKITBROWSER] Move the memory observer otherwise we have a dependency issue :-)
